Included docs. Updated default type to float32.

+++ kernel/carray.h
@@ -26,9 +26,9 @@
  * Currently working with shaped 2D, 1D and 0D.
  */
     typedef struct CArray {
-        double **  array2d;
-        double *   array1d;
-        double     array0d;
+        float **  array2d;
+        float *   array1d;
+        float     array0d;
     } CArray;
 
 /**
